Initial working version with appropriate split-adjustment logic for BB_spread strategy....

+++ strategies/xiQuant_strategies/bkup/xiquantFuncs.py
@@ -28,7 +28,7 @@ def normalize(value, mean, stdDev):
 	return float((value - mean) / stdDev)
 
 def slope(val1, val2):
-	return numpy.arctan((val1 - val2) / 2) * 180 / numpy.pi
+	return numpy.arctan(float((val1 - val2) / 2)) * 180 / numpy.pi
 
 def slopeForSeries(inpDS, lookbackWin):
 	if lookbackWin > 2:
